fit new API: failed :(

+++ evasion/clients/GameStrategy.py
@@ -33,16 +33,20 @@ class GameStrategy(object):
                 self.first_move = False
             else:
                 status = self.publisher.recv()
+                print "receive : %s" % status
                 status = json.loads(status)
                 hunter_pos = status["hunter"]
                 prey_pos = status["prey"]
-                walls = status["wall"]
+                walls = status["walls"]
                 time = status["time"]
                 hunter_direction = self._get_direction(self.hunter_prev_move, hunter_pos)
-                self._remove_walls(walls, hunter_pos)
+                # self._remove_walls(walls, hunter_pos)
                 wall_check = self._time_to_build_wall(hunter_direction, hunter_pos, prey_pos)
+
                 if time - self.last_wall_time > self.build_frequency and wall_check[0]:
                     self._build_entire_wall(wall_check[1])
+                    self.last_wall_time = time
+
                 self.hunter_prev_move = hunter_pos
                 self._send_moving_message()
 
@@ -72,7 +76,8 @@ class GameStrategy(object):
                     if NS_closest_id != -1:
                         remove_ids.append(NS_closest_id)
                     NS_closest_id = wall["id"]
-        self._delete_walls(remove_ids)
+        if len(remove_ids) >0:
+            self._delete_walls(remove_ids)
 
     def prey_strategy(self):
         """docstring for preyStrategy"""
@@ -109,7 +114,7 @@ class GameStrategy(object):
     def _time_to_build_wall(self, hunter_direction, hunter_pos, prey_pos):
         if hunter_direction == "SE":
             if prey_pos[1] - hunter_pos[1] < 4 and prey_pos[1] - hunter_pos[1] > 0:
-                return (True, "H")
+                return (True, "E")
             elif prey_pos[0] - hunter_pos[0] < 4 and prey_pos[0] - hunter_pos[0] > 0:
                 return (True, "V")
         elif hunter_direction == "NW":
@@ -204,7 +209,8 @@ class GameStrategy(object):
         self._send_to_server(message)
 
     def _build_entire_wall(self, direction):
-        message = {"command": "B", "direction": direction}
+        message = {"command": "B", "wall": {"length":100, "direction": direction}}
+        print "build wall : %s" % message
         self._send_to_server(message)
 
     def _delete_walls(self, wall_ids):
